maps/ipcache: Support IPv6 tunnel endpoint
This commit adds support for IPv6 tunnel endpoints in the ipcache map.

As a side effect, this commit adds support for IPv6 tunnel endpoints in
the "cilium-dbg bpf ipcache update" command.

Note the new IPv6.IsZero() function isn't used in this diff, but it's
called in the existing code for TunnelEndpoint.IsZero(). Since
TunnelEndpoint can now be IPv6 or IPv4, we need this new function to
handle the IPv6 case.

Note that there are up/downgrade consequences to this commit due to the
ipcache map layout changing and they are handled both in subsequent
commits, in a previous update to our Envoy fork, and in an upcoming
patch for v1.17 (for the downgrade path).

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>